Moved keypaths related to another file
